Generalize PathResolver to with a factory method for ClassPath instance.
PathResolver hard-coded `new JavaClassPath`. For LLVM, I'd like to
substitute a subclass instead. This patch turns PathResolver into an
abstract class, and references to it are replaced by JavaPathResolver.

It would be helpful if this change could be adopted upstream. It is
just a generalization, and contains no LLVM-specific functionality. I
tested that the resulting compiler can compile itself.