Use a portable ABI for fat references
Passing and returning structures in LLVM is bad for portability. As much
as possible we want the IR to be target independent so we pass fat
references as two separate arguments and return fat references by
returning the object pointer and storing the result vtable in a pointer
argument passed for that purpose.