Fix externalEquals forLLVM
Because BoxesRuntime for LLVM is implemented in scala there is no true
static and since JavaPlatform externalEquals method is named equals it
resolved to an OverloadedSymbol between the external equals and object
equals. Could have kept using equals and filter to get the right symbol
but it was much simpler to just change the name. We also need to fix up
the call type and load the BoxesRunTimeClass module.