Fix casting between interface types
A search of the instance's class info is required for all casts
to interface types. It should be possible to eliminate it in some
cases by being clever. For example, group the methods in an
interface vtable by the super-interfaces and let the resulting
vtable point to the middle of the sub-interface vtable.